Copying api_params.yaml file inside an exported API
