+++ import-export-cli/box/resources/init/api_params.tmpl
@@ -4,3 +4,8 @@ environments:{{- range $name, $elem := . }}
       production:
       sandbox:
 {{- end }}
+
+vcs:
+  import:
+    update: true
+    preserveProvider: true
