+++ import-export-cli/utils/fileIOUtils.go
@@ -25,6 +25,7 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"strings"
 
 	"gopkg.in/yaml.v2"
 )
@@ -356,4 +357,53 @@ func CreateZipFileFromProject(projectPath string, skipCleanup bool) (string, err
 		return projectPath, nil, cleanup
 	}
 	return projectPath, nil, nil
+}
+
+func GetTempCloneFromDirOrZip(file string) (string, error) {
+	fileIsDir := false
+	// create a temp directory
+	tmpDir, err := ioutil.TempDir("", "apim")
+	if err != nil {
+		_ = os.RemoveAll(tmpDir)
+		return "", err
+	}
+
+	if info, err := os.Stat(file); err == nil {
+		fileIsDir = info.IsDir()
+	} else {
+		return "", err
+	}
+	if fileIsDir {
+		// copy dir to a temp location
+		Logln(LogPrefixInfo+"Copying from", file, "to", tmpDir)
+		dest := filepath.Join(tmpDir, filepath.Base(file))
+		err = CopyDir(file, dest)
+		if err != nil {
+			return "", err
+		}
+		return dest, nil
+	} else {
+		// try to extract archive
+		Logln(LogPrefixInfo+"Extracting", file, "to", tmpDir)
+		finalPath, err := extractArchive(file, tmpDir)
+		if err != nil {
+			return "", err
+		}
+		return finalPath, nil
+	}
+}
+
+// extractArchive extracts the API and give the path.
+// In API Manager archive there is a directory in the root which contains the API
+// this function returns it appended to the destination path
+func extractArchive(src, dest string) (string, error) {
+	files, err := Unzip(src, dest)
+	if err != nil {
+		return "", err
+	}
+	if len(files) == 0 {
+		return "", fmt.Errorf("invalid API archive")
+	}
+	r := strings.TrimPrefix(files[0], src)
+	return filepath.Join(dest, strings.Split(filepath.Clean(r), string(os.PathSeparator))[0]), nil
 }
